@@ -11,16 +11,23 @@
     },
     "devDependencies": {
         "@popperjs/core": "^2.10.2",
+        "@vue/compiler-sfc": "^3.2.31",
+        "autoprefixer": "^10.4.4",
         "axios": "^0.25",
         "bootstrap": "^5.1.3",
+        "browser-sync": "^2.27.9",
+        "browser-sync-webpack-plugin": "^2.3.0",
         "laravel-mix": "^6.0.6",
         "lodash": "^4.17.19",
-        "postcss": "^8.1.14",
+        "postcss": "^8.4.12",
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.8",
-        "vue-template-compiler": "^2.6.12"
+        "tailwindcss": "^3.0.23",
+        "vue-loader": "^17.0.0"
+    },
+    "dependencies": {
+        "vue": "^3.2.31",
+        "vue-router": "^4.0.13"
     }
 }
