@@ -14,6 +14,18 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
-});
+// Route::post('login', [UserController::class, 'login']);
+// Route::post('register', [UserController::class, 'register']);
+// Route::post('logout', [UserController::class, 'logout'])->middleware('auth:sanctum');
+
+// Route::group(['prefix' => 'books', 'middleware' => 'auth:sanctum'], function () {
+//     Route::get('/', [BookController::class, 'index']);
+//     Route::post('add', [BookController::class, 'add']);
+//     Route::get('edit/{id}', [BookController::class, 'edit']);
+//     Route::post('update/{id}', [BookController::class, 'update']);
+//     Route::delete('delete/{id}', [BookController::class, 'delete']);
+// });
+
+// Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
+//     return $request->user();
+// });
